INS-876: relocate tests
