+++ application/extractor/base_test.go
@@ -15,3 +15,49 @@
  */
 
 package extractor
+
+import (
+	"testing"
+
+	"github.com/insolar/insolar/core"
+	"github.com/insolar/insolar/logicrunner/goplugin/foundation"
+	"github.com/stretchr/testify/require"
+)
+
+func TestStringResponse(t *testing.T) {
+	testValue := "test_string"
+
+	data, err := core.Serialize([]interface{}{testValue, nil})
+	require.NoError(t, err)
+
+	result, err := stringResponse(data)
+
+	require.NoError(t, err)
+	require.Equal(t, testValue, result)
+}
+
+func TestStringResponse_ErrorResponse(t *testing.T) {
+	testValue := "test_string"
+	contractErr := &foundation.Error{S: "Custom test error"}
+
+	data, err := core.Serialize([]interface{}{testValue, contractErr})
+	require.NoError(t, err)
+
+	result, err := stringResponse(data)
+
+	require.Contains(t, err.Error(), "Has error in response")
+	require.Contains(t, err.Error(), "Custom test error")
+	require.Equal(t, "", result)
+}
+
+func TestStringResponse_UnmarshalError(t *testing.T) {
+	testValue := "some_no_valid_data"
+
+	data, err := core.Serialize(testValue)
+	require.NoError(t, err)
+
+	result, err := stringResponse(data)
+
+	require.Contains(t, err.Error(), "Can't unmarshal")
+	require.Equal(t, "", result)
+}
